+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## x.x.x x-x-x
+
+### Payments
+* [Added] Support for Amazon Pay bindings.
+
 ## 23.23.0 2024-02-26
 ### PaymentSheet
 * [Added] Added support for [payment method configurations](https://docs.stripe.com/payments/multiple-payment-method-configs) when using the deferred intent integration path.